test: Consolidate scattered logic verification tests
